fix(_comp_compgen): prevent IFS from affecting `-- "$cur"`
The IFS separators specified by the options `-s SEP` or `-l` option is
intended to take effect in evaluating `compgen`, such as the word
splitting in `-W text`.

However, if we set IFS in a separate command from `compgen`, the
arguments specified to `compgen` would also be unexpectedly affected.
Here, ${_cur:+-- "$_cur"} is intended to be expanded to two words if
`_cur` is non-empty.  However, when the separators do not contain a
whitespace, it unexpectedly results in a single word.

IFS can be specified as a temporary environment of the `compgen`
command so that it does not affect the arguments of `compgen`.