refactor(ssh-add): use _comp_compgen_help
We also remove the backslash of '-\?' within this patch.  The
backslash was introduced in commit 60b8fabe, which seems to have tried
to work around the failglob failure.  Because we now directly specify
the arguments to _comp_compgen_help so that we do not have to evaluate
unquoted $2, we do not need to add the backslash anymore.  Note that
the backslash added to ssh-keygen in 60b8fabe was removed in commit
4dc00166 where _parse_help was switched to _parse_usage.

References:

https://github.com/scop/bash-completion/commit/60b8fabec499dbd56636548369ff3796e0d0d0fd
https://github.com/scop/bash-completion/commit/4dc0016625993487ed28ff0914ab8ebca47bd4e8