fix(7z): remove redundant `printf '%s\n' "..."`
Only one element is specified to printf, so it just appends a newline.
However, the newline is anyway removed by the word splitting performed
by -W.  We can safely remove the modifications by printf '%s\n'.

Before commit 1d3add44, the argument specified to `printf` was not
quoted so that each word splitted by IFS is printed to an independent
line.  However, as IFS is $'\n' in this context, words are obtained by
splitting the result with newlines, so outputting them by printf
'%s\n' does not have any effects.

References:
https://github.com/scop/bash-completion/commit/1d3add44563e48be7a6ec6cbd4cc8bb9e27d44f0#diff-efd1d1be37bc3e0c0b77e1a2d9ef05a7d9d7a09262b710adca23db163b18f387