build(check): run tests in `installcheck` rather than `check`
With `VPATH` builds, `completions/` symlinks are not created in the
source tree, but in the build one. This causes issues because the "main"
completion files the symlinks point to are not actually in the build
dir, and arranging them seems to be quite a chore with autotools it
seems. So the symlinks are dangling, and tests that depend on their
presence will fail.

Run tests against the installed tree instead; there we must have
everything in place, and testing the installed setup is a good thing
anyway.

Refs https://github.com/scop/bash-completion/pull/968