fix(_comp_count_args): perform optarg check also on $3
The check that unconditionally accepts a word by a pattern has been
added by commits 75ec298e and 3809d955.  However, even if the word
matches the specified pattern, if the word is an option argument of
the previous option, it should not be counted as an argument.  This
patch changes the behavior so that it does not treat the word matching
$3 as an argument when it is an option argument.  The current use case
of $3 is only by chmod, where $2 is unspecified, so the behavior is
unaffected by this change.