refactor(_comp_get_first_arg): return exit status
The current interface does not distinguish the case where no argument
is found and the case where the first argument is an empty string.
This patch returns whether the first argument is found through the
exit status.  This also modifies the callers so that they use the exit
status to check if the first argument is specified.