fix(dict): work around bash-4.3 ${v+"$@"} with custom IFS
Bash < 4.4 has a bug that all the elements are connected with
`${v+"$@"}` when IFS does not contain whitespaces.  This becomes
problem with `${dict_options+"${dict_options[@]}"}`.  One way to fix
it is to set IFS=$' \t\n'.  In this patch, we instead change the use
of the array so that we do not have to rely on the ${v+"$@"} pattern.