fix(_comp_count_args): exclude <>& from wordbreaks as _comp_initialize
In the current implementation, to count the number of arguments, we
reassemble `cword` and `words` by specified exclude chars (i.e.,
characters that should not be considered word breaks even though they
are in COMP_WORDBREAKS).  The reassembly by the top-level
`_comp_initialize` always specify `<>&` as exclude chars, but the
current implementaion of `_comp_count_args` do not specify them.  This
patch specifies `<>&` for `_comp_count_args` in accordance with
`_comp_initialize`.