fix(_known_hosts): use array for `options` (work around SC2178,SC2179)
shellcheck does not allow using a local scalar variable that has the
same variable name as an array variable used in another function
(SC2178,SC2179).  To use array `options` in another function, we
switch `options` in `_known_hosts` to an array variable (or otherwise,
we will need to place disable=SC2178,SC2179 to every line using the
variable as a scalar).