fix(completions/*): scan until cword instead of ${#words[@]}-1
We find in the codebase many loops scanning the arguments excluding
the last word (i.e., the loop condition is `i < ${#words[@]} - 1`).

The earliest versions of this type of loops were introduced in commit
e8d1508ce in the form of `${#COMP_WORDS}-1` when the completions for
the `apt-*` commands are introduced.  There is no explanation about
the reason that the last word is excluded. [ Note: `${#COMP_WORDS}`
counting characters instead of words is clearly wrong but later fixed
to be `${#COMP_WORDS[@]}` in commit 57437d515. ]

Those are likely to be intended to exclude the word we currently try
to complete.  If that is the case, we should instead exclude all the
words at the current word and later words (i.e., the loop condition is
`i < cword`).  The value of `cword` matches `${#words[@]} - 1` in a
typical case where the tab completion is attempted at the end of the
command line, but this assumption breaks down when the user attempts a
completion in the middle of the command line.