fix(qdbus): remove cur and all the rest arguments
The current codebase removes the last word when `cur` has non-empty
value.  This piece of code was originally introduced in commit
1d74adc29.  The original code before this commit tried to remove the
suffix ` $cur` from ${COMP_WORDS[*]}.  This means that the last word
was assumed in the commit to be the current word, which is not the
case when the user attempts the completion in the middle of the
command line.  In this patch, only the words before the current word
is used, (i.e., the current word (at the index of `cword`) and all the
rest words are removed).