fix(sysbench): fix extraction of test-name option
The `=` seems to be missing in the pattern.  This pattern was in this
form from the beginning when the sysbench completion is introduced in
commit 77b78b372.  With the current implementation, `--test`,
`--testxxxx`, etc. can also match, and $test becomes `--test`,
`--testxxxx`, etc.  This results in the later evaluation of
`--test=--test` or `--test=--testxxxx`, which doesn't seem to be
correct.  Therefore, `=` is suspected to be missing.