refactor(tar): remove `_comp_cmd_tar__last_char`
The last character of the variable can be obtained by simply ${var:
-1} from bash-2.0, so there is no need to use a helper function to
calculate the offset from the variable length.

The function is originally introduced in commit 8b23c84cc as a
function named `__tar_last_char` along with many other functions, but
there does not seem to be specific reason for using the function to
get the last character of variables.