feat(_comp_compgen): support `-P prefix` with auto-adjusted `cur`
The Bash builtin `compgen -P prefix ... -- "$cur"` prepends the prefix
AFTER filtering completions using `$cur`.  However, we usually want to
filter completions with "$cur" starting with the prefix.

To properly handle such a situation, one first needs to check if the
current content of "$cur" is compatible.  Then, one needs to modify
$cur to remove the prefix part, generate completions, and prepends the
prefix to the generated completions.  This pattern is used frequently
in the codebase, so it is good to handle it within `_comp_compgen`.

This patch implements the option `-P` of `_comp_compgen`.  When a
non-empty string is specified to the `-P` option, it performs the
necessary operations: the check and adjustment of $cur, the proper
filtering by the prefix string, and prepending of the prefix string.

Co-authored-by: Ville Skyttä <ville.skytta@iki.fi>