fix(perl): fix ineffective condition in an awk script
The condition `$NF=2` in the awk script actually assigns a value `2`
to the last field instead of comparing it to `2` and always evaluates
to true.  I suspect it to be `NF >= 2`.

This could possibly be intended as `NF == 2` or `$NF == 2`, but
neither seems to work correctly.  The input contains the names of man
entries of the form ` perlxx <description>`, but the description is
usually given by more than one words, so restricting them by `NF == 2`
would be unreasonable.  On the other hand, there are no entries ending
with the word `2` so `$NF == 2` would produce no results.  This
condition was introduced from the beginning when the related code was
introduced in commit 4254f3a4.  There does not seem to be any hints on
the background of `$NF=2`.