fix(_mock,rpm): avoid icase flag `s/reg/rep/i` of GNU sed
Flag i for the case-insensitive matching is a GNU extension and thus
should be avoided.

The mock completion contains s/.../.../i while the rpm completion also
contains a similar code but without flag i.  The related sed code in
the rpm completion was introduced in commit 14c9f5b23c (2009-05-23),
and that in the mock completion was introduced in commit 508a88e6c0
(2009-05-31).  The former did not have flag i while the latter had
flag i from the beginning.

In this patch, we assume the case-insensitive flag was intended for
both cases and update the regular expression so that it matches in a
case-insensitive manner also in the rpm completion.