style(_comp_awk): prefer `awk -F <char>` to `awk -F<char>`
We have been using the connected form of the `-F` option to the `awk`
command because Solaris awk does not support the separate form of the
option argument `-F <char>`.  Nevertheless, there are some cases of
the separate form in the codespace, though the majority is in the
connected form.  Since we now switch to `_comp_awk` for POSIX awk, we
unify them to the separate form `-F <char>`.