fix(_comp_delimited): treat delimiter as a fixed string
The right-hand side of == in the conditional commands [[ ]] are
treated as a glob pattern.  In the other occurrences of $delimiter,
they are quoted as "$delimiter", which means that $delimiter treated
as a fixed string.  However, only in this line, $delimiter is treated
as a glob pattern.  In this patch, we quote it properly the same as in
the other occurrences of $delimiter.