fix: use compgen -P prefix -W '"${arr[@]}"' against patsub_replacement
To add a prefix to every element in an array, we have used the
following construct:

  arr=("${arr[@]/#/$prefix}")

However, this can be broken when `shopt -s patsub_replacement` (bash
>= 5.2) is turned on.  In bash >= 5.2, patsub_replacement is enabled
by default.  In particular, the characters `&` contained in $prefix,
if any, will be replaced with the matched string.  To avoid the
unexpected patsub_replacement, one may think about quoting the
replacement as

  arr=("${arr[@]/#/"$prefix"}")

However, this has another problem in bash < 4.3 or when `shopt -s
compat42` is turned on.  In such a case, the inner double quotations
are treated literally so that the `PREFIX` instead of ``"PREFIX"` is
prefixed to elements.  To avoid this situation the outer double
quotations might be removed, but this has even another problem of the
pathname expansions and IFS.  We here instead use _comp_compgen:

  _comp_compgen -Rv arr -- -P "$prefix" -W '"${arr[@]}"'

This patch also contains similar cases for a suffix.