fix(ri): avoid interference on existing COMPREPLY
The current implementation of `_comp_cmd_ri__methods` prefixes $prefix
to the generated method names.  However, this prefixing also affects
the already generated completions in COMPREPLY because it first
generates the method names by appending them to COMPRELY and then
modifies all the elements of COMPREPLY.

In the current usage, it did not produce the problem because prefix is
empty when there are existing elements in COMPREPLY, but it's
accidentally.  Before refactoring the code, we want to resolve this
dependency on the subtle condition.