fix(available_interfaces): fix regression of unwanted trailing colons
This is a fix for regression introduced in commit b6035350 and
reported in GitHub #1129 [1].  The trailing colons of the generated
interface names were removed before, but not they are not removed.

[1] https://github.com/scop/bash-completion/issues/1129

In addition, generated words can have the form `veth0@veth1`, where we
want only the first part `veth0` not containing any punctuation
characters.

In this patch, we remove any [[:punct:]] and all the later characters
in the generated words.