fix: use `pathcmd=$(type -P -- "$1")` for arbitrary cmds
If the command name starts with `-`, `type` is confused and tries to
parse it as an option to `type`.  In the current setup, the option
starting with `-` does not happen because we do not set up the
completion settings for such command names, but it is better to put
them on a safe side.