build(pkgutil): don't install on non-Solaris platforms
In commit 0e3a17df7eac79ff386891945a951a23134373b0, a short-circuit was added
that causes the completion to return 1 if the host OS is not solaris.

On Gentoo, this triggers a warning:

```
 * QA Notice: Problems with installed bash completions were found:
 *
 *      pkgutil: does not define any completions (failed to source?).
 *
 * For more details on installing bash-completions, please see:
 * https://wiki.gentoo.org/wiki/Bash/Installing_completion_files
```

The purpose of the check is to find broken completions that need to be fixed by
their respective upstreams, and in this case the completion is "intentionally
broken". It seems straightforward to respect that at an earlier level, which
both avoids the QA check and ensures that unneeded files don't get installed on
any distro.

And... it also means that macOS users can ship their own pkgutil completion
without worrying about file clashes with the bash-completion project. The
entire reason for the short-circuit was because there's a *different* pkgutil
there.

Fixes: 0e3a17df7eac79ff386891945a951a23134373b0