feat(aptitude): update sub-commands
The following commands are listed in the manpage and were missing from
the completion file:

    build-depends
    extract-cache-subset
    help

Whereas the following commands are not listed in the manpage nor in the
documentation:

    dist-upgrade
    upgrade

Moreover, build-depends and build-deps are aliases, and bash-completion
guidelines (CONTRIBUTING.md) recommend using the longer version.

This patch updates the list and sorts it by-name.

Suggested-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
Suggested-by: Koichi Murase <myoga.murase@gmail.com>