fix(ip): Don't offer neighbour, only neighbor
The previous state led to `ip neigh<TAB>` leading to this:
```
$ ip neighbo
neighbor   neighbour
```
choosing one to complete removes the need to choose between options that
do the same thing, aligning with the CONTRIBUTING.md guidelines.

I chose `neighbor` because it's shorter, but it can be changed if someone
wants to.