refactor(ssh): remove outdated disable=SC2089,SC2090
These shellcheck directives were added in commit c906aeb (2020-04-10),
but the latest version of shellcheck-0.10.0 does not seem to warn
about SC2089 and SC2090 on these lines for the current version of
bash_completion.

These warnings seem to be issued when unquoted
$_comp_cmd_scp__path_esc is used, such as

    echo $_comp_cmd_scp__path_esc

We have already resolved those unquoted variable expansions, so the
warning should no longer happen.  In this patch, we remove those
outdated shellcheck directives.