fix(rsync,scp): remove file-type marks from "ls -F" properly
The type-classifier mark for named pipes (|) and sockets (=) were not
properly removed.  These are escaped by a backslash, so the backslash
remains if we only remove the character of the type classifier.  In
this patch, we remove the backslash altogether for characters that are
included in _comp_cmd_scp__path_esc.