fix(_comp_dequote): set the literal value to REPLY as a fallback
It is more useful to set the literal value of $1 to REPLY even when $1
is an unsafe word.  When the caller wants to use the result only when
REPLY is a safe string and suceceds without failglob, the caller can
reference the exit status $?.  When the caller wants to use any
string, it is more useful if REPLY contains a fallback value.

This patch changes the behavior so that REPLY is set to be the literal
value of $1 even when an unsafe value is specified to $1.  We also
change the exit status to be 0 only when at least one word is
generated by a safe string.