fix(java,rsync,ssh): complete syntactically incomplete cur
https://github.com/scop/bash-completion/issues/1255#issuecomment-2437499379

`cur` is the word that is currently input by the user and can be
syntactically incomplete when the completion is requested.  For
example, it is typical to attempt a completion with an opening
quotation `'` but without a closing quotation:

$ javadoc 'a[tab]

Or, if there are two candidates `a\ b.txt` and `a\\\ c.txt`, the first
attempt of the completion would insert the common part `a\`, where the
escape target of `\` is missing.  To handle these cases, in dequoting
values coming from `cur`, we need to deal with the cases with
incomplete values.

This patch adds a new utility `_comp_dequote_incomplete` and replace
`_comp_dequote` with it everywhere `_comp_dequote` is used for values
coming from `cur`.

Co-authored-by: Yedaya Katsman <yedaya.ka@gmail.com>