fix(mdtool): fix usage of `compgen -G`
The original code (which is commented out and marked as TODO) tried to
perform the filename generation by specifying both `-f` and `-G`, and
also specifying `-G` multiple times.  This has the following issues:

* The completion specification of `-f` and `-G` are independent, so
  this code is going to generate all filenames with `-f`. Then, it
  additionally generates the filenames matching the specified pattern.
* The `-G pat` can only be specified once, or the latter `-G pat`
  specification will overwrite the previous specification.  We should
  combine multiple patterns into one.

Note: This code stub was originally introduced in commit 73225ea4.

In this patch, we remove `-f` and merge multiple `-G`.