fix(cd): fix the pattern matching
This code appears to try to suffix "/" when the unique completion of a
directory name does not end with "/".  However, the test compares the
completion with the quoted `"*/"`, which does not work as expected.
If the generated directory name already ends with "/" (though it is
unclear whether it may happen except for "/"), this code would result
in double slashes such as "dir//".

This double-quoted pattern is present from the beginning when this
code was introduced in commit a1969bdd57.

This patch removes the double quoting so that the directory name is
correctly matched with the pattern `*/`.