refactor(cd): factorize generator for CDPATH
To apply the adjustment of suffixing "/" only to the completions
generated based on CDPATH, we isolate the CDPATH completion code in a
separate generator.  This is needed because we want to generate
cdable_vars completions into COMPREPLY, but we do not want to apply
sufixing to the cdable_vars completions.