test(interfaces): Test for missing ip or ifconfig
We have a fallback to `ifconfig` if `ip` doesn't exist, but it wasn't
tested anywhere that does have `ip` (pretty sure that's all of our current
distro tests).

Introduce a parameterized fixture that "hides" `ip` or `ifconfig` and,
use it in tests.
Also introduce another test that actually makes sure that we find the
correct interfaces.

Follow up to https://github.com/scop/bash-completion/pull/1090