refactor: move files to separate bundled/external completions
This implements the separation of bundled/external completion files
suggested in Ref. [1].

To avoid unnecessary renamings of the files in the Git history,
suffixing `.bash` [2] and separation of fallback completion [3] are
also processed at once.

All completion files with a prefix underscore were moved into
`completions-fallback` (including 3rd-party fallback completions). The
symbolic links `hd -> _hexdump` and `ncal -> _cal` are also created in
`completions-fallback` because the target files are fallback
completions (with an underscore `_` as `_hexdump` and `_cal`). The
other files are moved into `completions-core`.

[1] \
https://github.com/scop/bash-completion/discussions/1329#discussioncomment-12163012.
[2] \
https://github.com/scop/bash-completion/discussions/1329#discussioncomment-12163019
[3] \
https://github.com/scop/bash-completion/discussions/1329#discussioncomment-12163025