refactor(test/fallback): symlink "completions-fallback" on the fly
Background: `test/fallback/completions` has been used to override the
precedence of the completion lookup of fallback completions.  With the
traditional search order of completion files, if the system where the
test is run provides a completion file, that completion file is loaded
instead of a fallback completion file in the bash-completion tree.
However, we want to test our in-tree fallback completion.  To
correctly select our in-tree fallback completion, we set up
`test/fallback/completions` and specified `test/fallback` to
BASH_COMPLETION_USER_DIR.

With the new search order, completions in the test environment will be
used instead of the in-tree completions even for non-fallback
completions.  To change the ordering of the completions, one would
need to make symbolic links for all completions in
`test/fallback/completions`.  However, with such a setup, all files
are duplicated as hard links in the tarball created by `make dist`
(because Autoconf creates a tarball with the symbolic links fully
resolved to hard links).

In this patch,

* We give up providing pre-defined `test/fallback/completions` in the
  Git tree.  We remove the directory `test/fallback`.

* We instead prepare BASH_COMPLETINO_USER_DIR directories on the fly,
  where symbolic links "completions" to "completions-core" and
  "completions-fallback" are created, as a pytest fixture
  "test_session_tmpdir".

* HISTFILE has been created separately inside the fixture "bash", but
  we now create HISTFILE inside "test_session_tmpdir" shared by
  different Bash processes.