test(TestCompLoad.fixture_dir): specify the scope
Currently, the fixture directory is created for every test because the
scope of the fixture is not specified.  Since the contents of the
fixture directory is fixed and each test does not modify its contents,
it should be safe to share the fixture directory.

In this patch, we specify the scope of the fixture to be "class", so
that the fixture directory is shared by tests.